+++ CHANGELOG.md
@@ -9,6 +9,7 @@ This is a log of major user-visible changes in each phpMyFAQ release.
 ### phpMyFAQ v3.1.5 - 2022-05-
 
 - added compatibility with Elasticsearch v8+ (Thorsten)
+- added trust of self-signed certificates with MS SQL (Thorsten)
 - fixed minor bugs (Thorsten)
 
 ### phpMyFAQ v3.1.4 - 2022-04-25